refactor: update path and imports
